Move recent file additions to initial view handler
