+++ src/main/kotlin/com/mituuz/fuzzier/Fuzzier.kt
@@ -295,33 +295,9 @@ open class Fuzzier : FuzzyAction() {
             }
         }
         if (fuzzyMatchContainer != null) {
-            addFileToRecentlySearchedFiles(fuzzyMatchContainer)
+            InitialViewHandler.addFileToRecentlySearchedFiles(fuzzyMatchContainer, fuzzierSettingsService)
         }
         popup?.cancel()
-    }
-
-    fun addFileToRecentlySearchedFiles(fuzzyMatchContainer: FuzzyMatchContainer) {
-        var listModel: DefaultListModel<FuzzyMatchContainer>? = fuzzierSettingsService.state.recentlySearchedFiles
-
-        if (listModel == null) {
-            listModel = DefaultListModel<FuzzyMatchContainer>()
-            fuzzierSettingsService.state.recentlySearchedFiles = listModel
-        }
-
-        var i = 0
-        while (i < listModel.size) {
-            if (listModel.get(i).filePath == fuzzyMatchContainer.filePath) {
-                listModel.remove(i)
-            } else {
-                i++
-            }
-        }
-
-        while (listModel.size > fuzzierSettingsService.state.fileListLimit - 1) {
-            listModel.remove(listModel.size - 1)
-        }
-
-        listModel.addElement(fuzzyMatchContainer)
     }
 
     private fun createListeners(project: Project) {
